Add files via upload
Business Objective:
The objective is to device a machine learning model to predict the total
payment for each brand. The objective is to device a model to group the products
based on their brands and the locality where they have been sold and prioritize the
products based on the sales. Our objective is to find out which brands products is
getting high sales.